requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
PyPDF2
python-docx
faiss-cpu
python-dotenv
streamlit-lottie
requests
nest_asyncio
